+++ Assets/Scripts/Player/Weapon.cs
@@ -2,7 +2,7 @@
 [System.Serializable]
 public class Weapon  {
     
-    public int DamagePerShot = 20;
+    public float DamagePerShot = 20;
     public float TimeBetweenBullets = 0.15f;
     public float Range = 100f;
